Do a little memory management and improve the color rendering code

+++ Source/Engine/Renderer/Renderer.cpp
@@ -29,7 +29,7 @@ namespace ane {
 		SDL_RenderPresent(this->renderer);
 	}
 
-	void Renderer::SetColor(int r, int g, int b, int a) {
+	void Renderer::SetColor(uint8_t r, uint8_t g, uint8_t b, uint8_t a) {
 		SDL_SetRenderDrawColor(this->renderer, r, g, b, a);
 	}
